Rollback to doctrin/common:<3.0 and doctrine/doctrine-bundle:<2.0
In doctrine/common:^3.0, there will be a conflict with FOSUserBundle that is currently working because the ObjectManager class will change.
Also this class is used by the actual project in the construct method of the UserManager (extended from FOSUserBundle). Since FOSUserBundle didn't have upgraded to that version, we need to wait for it.